backend/requirements.txt
# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0


# LangChain - versões compatíveis
langchain==0.1.20
langchain-openai==0.1.7
langchain-community==0.0.38
langchain-text-splitters==0.0.1
langchain-core==0.1.52
langsmith==0.1.55

# Vector DB e ML
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.0

# PDF processing
PyPDF2==3.0.1

# Utilities adicionais
requests==2.31.0
aiofiles==23.2.1
pypdf==3.17.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0